fix: move project to Visual Studio 2022
It now uses build tools v143 and Windows SDK 10.0.

+++ src/RenderEngine/DirectX9/LightFieldInstance.cpp
@@ -147,7 +147,7 @@ LightFieldInstance::LightFieldInstance(const LightFieldSource& source, RenderObj
 
     // Determine XY rotation of attached bone as initial rotation
     const Vector3 dir = Vector4(1,0,0,0) * m_object.GetModel().GetBone(m_bone).absTransform;
-    m_rotation = atan2(dir.y, dir.x);
+    m_rotation = atan2f(dir.y, dir.x);
 }
 
 LightFieldInstance::~LightFieldInstance()
